fix: update address count to return an int

+++ app/src/main/java/com/littlelemon/application/address/data/local/AddressLocalDataSource.kt
@@ -14,5 +14,5 @@ interface AddressLocalDataSource {
 
     suspend fun saveAddresses(addresses: List<AddressEntity>)
 
-    suspend fun getAddressCount(): Long
+    suspend fun getAddressCount(): Int
 }
