+++ app/src/main/java/com/littlelemon/application/address/domain/usecase/GetAddressCountUseCase.kt
@@ -3,5 +3,5 @@ package com.littlelemon.application.address.domain.usecase
 import com.littlelemon.application.address.domain.AddressRepository
 
 class GetAddressCountUseCase(private val repository: AddressRepository) {
-    suspend operator fun invoke(): Long = repository.getAddressCount()
+    suspend operator fun invoke(): Int = repository.getAddressCount()
 }
